docker compose files and config data moved separate project
