+++ pom.xml
@@ -23,6 +23,12 @@
 
 	<dependencyManagement>
 		<dependencies>
+			<!-- LOCAL-->
+			<dependency>
+				<groupId>com.javadaily.demo</groupId>
+				<artifactId>app-config-data</artifactId>
+				<version>${project.version}</version>
+			</dependency>
 			<dependency>
 				<groupId>org.springframework.boot</groupId>
 				<artifactId>spring-boot-starter</artifactId>
@@ -79,5 +85,6 @@
 
 	<modules>
 		<module>twitter-to-kafka-service</module>
+		<module>app-config-data</module>
 	</modules>
 </project>
