+++ twitter-to-kafka-service/pom.xml
@@ -11,6 +11,10 @@
 
 	<dependencies>
 		<dependency>
+			<groupId>com.javadaily.demo</groupId>
+			<artifactId>app-config-data</artifactId>
+		</dependency>
+		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter</artifactId>
 		</dependency>
@@ -27,16 +31,12 @@
 		</dependency>
 
 		<dependency>
-			<groupId>org.projectlombok</groupId>
-			<artifactId>lombok</artifactId>
-			<scope>provided</scope>
-		</dependency>
-
-		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-configuration-processor</artifactId>
 			<optional>true</optional>
 		</dependency>
+
+
 	</dependencies>
 
 	<build>
