+++ twitter-to-kafka-service/src/main/java/com/javadaily/demo/twitter/to/kafka/service/runner/impl/MockKafkaStreamRunner.java
@@ -13,7 +13,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.stereotype.Component;
 
-import com.javadaily.demo.twitter.to.kafka.service.config.TwitterToKafkaServiceConfiguration;
+import com.javadaily.demo.config.TwitterToKafkaServiceConfiguration;
 import com.javadaily.demo.twitter.to.kafka.service.exception.TwitterToKafkaServiceException;
 import com.javadaily.demo.twitter.to.kafka.service.listener.TwitterKafkaStatusListener;
 import com.javadaily.demo.twitter.to.kafka.service.runner.StreamRunner;
